{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name ="{{cookiecutter.package_name}}"
version = "{{cookiecutter.version}}"
description = "{{cookiecutter.project_short_description}}"
authors = ["{{cookiecutter.author_name}} <{{cookiecutter.author_email}}>"]


[tool.poetry.dependencies]
python = "^3.7"
structlog = "^20.2.0"
python-dotenv = "^0.15.0"
pydantic = "^1.7.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"
docformatter = "^1.4"
pylint = "^2.6.0"
flake8 = "3.8.4"
mypy = "^0.800"
isort = "^5.7.0"
bandit = "^1.7.0"
black = "20.8b1"
jupyter = "^1.0.0"
matplotlib = "^3.3.4"
sphinx-rtd-theme = "^0.5.2"
sphinx-pydantic = "^0.1.1"
sphinx-autodoc-typehints = "^1.12.0"
seed-intersphinx-mapping = "^0.5.1"
coverage = "^5.5"

[tool.black]
line-length = 79

[build-system]
requires = [
    "setuptools>=42",
    "wheel",
]
build-backend = "setuptools.build_meta"
# This could be using poetry build backend 
#[build-system]
#requires = ["poetry_core>=1.0.0"]
#build-backend = "poetry.core.masonry.api"